add els config
