+++ bootstrap/src/main/java/org/hesperides/HesperidesSpringApplication.java
@@ -3,6 +3,9 @@ package org.hesperides;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.data.elasticsearch.ElasticsearchAutoConfiguration;
+import org.springframework.boot.autoconfigure.data.elasticsearch.ElasticsearchDataAutoConfiguration;
+import org.springframework.boot.autoconfigure.data.elasticsearch.ElasticsearchRepositoriesAutoConfiguration;
 import org.springframework.boot.autoconfigure.data.mongo.MongoDataAutoConfiguration;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceTransactionManagerAutoConfiguration;
@@ -14,6 +17,9 @@ import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfigurat
         exclude = {
                 DataSourceAutoConfiguration.class,
                 DataSourceTransactionManagerAutoConfiguration.class,
+                ElasticsearchAutoConfiguration.class,
+                ElasticsearchDataAutoConfiguration.class,
+                ElasticsearchRepositoriesAutoConfiguration.class,
                 HibernateJpaAutoConfiguration.class,
                 MongoAutoConfiguration.class,
                 MongoDataAutoConfiguration.class
